fix: fixed json deserializers and idade param to train our neural network

+++ src/main/java/br/unisc/tuberculosis_rna/command/TuberculosisRNACommand.java
@@ -4,5 +4,5 @@ public record TuberculosisRNACommand(String encryptedFilePath,
                                      int numCamadas,
                                      int tamCamada,
                                      double taxaAprendizado,
-                                     int margemErro,
+                                     double margemErro,
                                      int numInteracoes) {}
